requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
PyPDF2
sentence-transformers
faiss-cpu
python-multipart
google-generativeai
groq
duckduckgo-search
pytest
httpx
itsdangerous>=2.0
pydantic-settings>=2.0
PyMuPDF>=1.20
grpcio>=1.50
langdetect>=1.0
pycountry>=22.3.5
openai>=1.0
aiosqlite>=0.17
python-jose>=3.3
pytesseract>=0.3.10
openai-whisper>=202311
authlib>=1.2
cryptography==41.0.5
